CFP-1754 Improve tests for pagination and adapt code accordingl

+++ internal/app/cloudgontroller/api/v3/controllers/common/pagination.go
@@ -2,7 +2,7 @@ package common
 
 type PaginationParams struct {
 	Page    int `query:"page" validate:"gte=1"`
-	PerPage int `query:"per_page" validate:"gte=1,lte=5000"`
+	PerPage uint16 `query:"per_page" validate:"gte=1,lte=5000"`
 }
 
 func DefaultPagination() PaginationParams {
